refactor: contract setup logic
